backend/requirements.txt
fastapi 
passlib[bcrypt]
pydantic
sqlalchemy
psycopg2-binary 
uvicorn
python-jose
python-multipart
pydantic[email]
python-dotenv
requests
